Free bound parameters after Execute() is called
Previously the paremeters had the oppurtunity to be garbage collected
before Execute() was run because they were locally scoped in the bind()
function. Now we create CStrings and pass pointers to them to
OCIBindByPos and return these parameters to the caller to be free'd
after the Execute method is called and parameters used.

Also refactored nil type check to be part of the general type switch.

Signed-off-by: Sheena McCoy <sp.mccoy@modcloth.com>