Enabling real external authentication
If you want to use external authentication (os level or LDAP), you need
to tell the API to exactly do that.

This patch tries to implement this the common way of not having a
username, a password and a host to connect an. In that case, the
external authentication is triggered and the credentials are changed
in the OCI8 C API to external (OCI_CRED_EXT).